arenaflow/backend/pyproject.toml
[tool.poetry]
name = "arenaflow-backend"
version = "0.1.0"
description = "Smart Venue Intelligence Platform API"
authors = ["Admin <admin@arenaflow.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.103.0"}
uvicorn = {extras = ["standard"], version = "^0.23.2"}
pydantic = "^2.3.0"
pydantic-settings = "^2.0.3"
sqlalchemy = "^2.0.20"
asyncpg = "^0.28.0"
alembic = "^1.12.0"
upstash-redis = "^1.1.0"
firebase-admin = "^6.2.0"
prophet = "^1.1.4"
scikit-learn = "^1.3.0"
pandas = "^2.1.0"
numpy = "^1.26.0"
googlemaps = "^4.10.0"
google-cloud-translate = "^3.12.0"
slowapi = "^0.1.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
arq = "^0.25.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.0.287"
black = "^23.9.1"
mypy = "^1.5.1"
pytest = "^7.4.2"
pytest-asyncio = "^0.21.1"
faker = "^19.6.0"
httpx = "^0.25.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
